+++ lib/core/base-provider.js
@@ -1,7 +1,7 @@
 const providers = require("./providers");
 const path = require("path");
 const ncConfig = require(path.join(process.cwd()), ".nc.config");
-const validProviders = require("./provider_list");
+const validProviders = require("./providers_list");
 
 class Provider {
   /**
