+++ lib/core/providers_list.js
